fix: correct progress bar timing and rename chunk vars to batch
Move progressBar.tick() after the import completes so progress reflects
actual work done. Rename chunk/chunkSize to batch/batchSize to
distinguish from content chunking. Remove unused totalProcessed.
